add option to choose whether to create qrCode
