+++ src/qrcode.ts
@@ -62,3 +62,20 @@ export const settings = {
 export function create(overrides: any = {}) {
   return new (QRCodeStyling as any)(deepmerge(settings, overrides))
 }
+
+export async function generateQRCode(data: string) {
+  let qrCode = create({})
+  qrCode.update({
+    data,
+  })
+  let blob = await qrCode.getRawData('svg')
+  let qrCodeData: string = await new Promise((resolve, reject) => {
+    let reader = new FileReader()
+    reader.readAsDataURL(blob)
+    reader.onloadend = function () {
+      resolve(reader.result as any)
+    }
+  })
+
+  return qrCodeData
+}
